PC
I generated a game where when you click on the text "Refresh Me". It generates a random dice numbers for the two dices. If PlayerOne's number is more then Player 1 wins otherwise Player 2 wins or it draws. On Clicking the text which replaced the "Refresh Me" it generates next numbers and as game continues..............Happy Gaming!!!!!!!!!!